requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=1.0.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
streamlit>=1.29.0
openai>=1.0.0

# Enhanced features (optional but recommended)
pypdf>=3.0.0
unstructured>=0.10.0
python-docx>=1.0.0
rank-bm25>=0.2.2

